minor improvements consistency and concision

+++ outlier_states/school_spending_outliers.py
@@ -4,10 +4,10 @@ filename = "Public School K-12 Education Spending Per Pupil by State 2014.txt"
 
 states = pandas.read_csv(filename, sep='\t', usecols=range(2), skiprows=[1])
 
-mean = int(states["Total"].mean())
-stdev = int(states["Total"].std())
+mean = states["Total"].mean()
+stdev = states["Total"].std()
 
-outlier = abs(states['Total'] - mean) > stdev
+outlier = abs(states["Total"] - mean) > stdev
 outlier_states = states[outlier]
 
 print(states)
